+++ lib/src/clixon_netconf_lib.c
@@ -1375,7 +1375,7 @@ netconf_minmax_elements_xml(cxobj **xret,
 	goto done;
     }
     if (xml_parent(xp)){ /* Dont include root, eg <config> */
-	if (xml2xpath(xp, &path) < 0)
+	if (xml2xpath(xp, NULL, &path) < 0)
 	    goto done;
 	if (path)
 	    cprintf(cb, "%s", path);